Code/requirements.txt
pandas
numpy
xlrd
factor_analyzer
matplotlib
PyYAML
Cython
supplychainpy==0.0.5